api restructured for clarity
