+++ backend/flask_api/test.py
@@ -4,7 +4,7 @@ BASE = "http://127.0.0.1:5000"
 
 response = requests.put(
     BASE + "/user",
-    json={"action": "mod_tier", "user_name": "bob", "tier": "pro"},
+    json={"action": "change", "user_name": "bob", "pwd": "bobsmith", "new_pwd": "smth2"},
     timeout=10,
 )
 print(response.json())
